Substitutes a forked repo for mybinder PDSH link
The Mybinder.org link for Jake van der Plas' Python Data Science Handbook
(PDSH) is currently broken (and has been for a while: jakevdp/PythonDataScienceHandbook#108).
The root of the problem is an incompatibility between the version of
scikit-image stated in the requirements.txt and the 3.6+ version of Python.
Upgrading the scikit-image requirement has been stalled for > 6 months
(see jakevdp/PythonDataScienceHandbook#117). The alternative is to force the
Python runtime to Python 3.5, which is what one fork did:
baldwint/PythonDataScienceHandbook@b23349050e2eaeefabedcfad753d00f8c89903d3

Here we're commenting out the link that we'd really want but that's broken
(bad for learners), and instead use the link that uses the aforementioned
fork.

This obviously has to be a temporary fix only. Currently there is no sign
that the problem will be fixed in the original repo by merging
jakevdp/PythonDataScienceHandbook#117. One more long-term reliable solution
to this could be to fork the original repo here and apply the same (trivial)
fix as in the aforementioned fork. Alternatively, we may consider switching
to another example.