Substitutes the Data Exploration notebook for mybinder PDSH link
The Mybinder.org link for Jake van der Plas' Python Data Science Handbook
(PDSH) is currently broken (and has been for a while: jakevdp/PythonDataScienceHandbook#108).
The root of the problem is an incompatibility between the version of
scikit-image stated in the requirements.txt and the 3.6+ version of Python.
Upgrading the scikit-image requirement has been stalled for > 6 months
(see jakevdp/PythonDataScienceHandbook#117). The alternative is to force the
Python runtime to Python 3.5, which is what one fork did:
baldwint/PythonDataScienceHandbook@b23349050e2eaeefabedcfad753d00f8c89903d3

We could switch to this fork, but this obviously would have to be a temporary
fix only, and would almost be guaranteed to break again in the future.
Currently there is no sign that the problem will be fixed in the original
repo by merging jakevdp/PythonDataScienceHandbook#117.

So instead of using a temporary fix and/or waiting for whenever the original
repo will be fixed, this proposal switches to a notebooke "of our own",
which has the advantages of having only few and relatively common dependencies,
and being one that the learners just worked through, and now get to see a
live-shared version of.